requirements.txt
absl-py==2.3.1
ai-edge-litert==2.1.0
amqp==5.3.1
apt-listchanges==4.8
arrow==1.3.0
asgiref==3.8.1
astunparse==1.6.3
attrs==25.3.0
autocommand==2.2.2
av==14.2.0
babel==2.17.0
backports.strenum==1.2.8
bcrypt==4.2.0
beautifulsoup4==4.13.4
billiard==4.2.4
blinker==1.9.0
celery==5.3.6
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloud-init==25.2
colorzero==2.0
configobj==5.0.9
cryptography==43.0.0
cssselect==1.3.0
cupshelpers==1.0
dbus-python==1.4.0
distro==1.9.0
Django==5.0.1
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
Flask==3.1.1
flatbuffers==20181003210633
fqdn==1.5.1
freetype-py==2.5.1
gast==0.7.0
google-pasta==0.2.0
gpiod==2.2.0
gpiozero==2.0.1
grpcio==1.76.0
gunicorn==21.2.0
h5py==3.15.1
html5lib-modern==1.2
hx711==1.1.2.3
idna==3.10
inflect==7.3.1
isoduration==20.11.0
itsdangerous==2.2.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jaraco.text==4.0.0
Jinja2==3.1.6
jsonpatch==1.32
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.12.1
keras==3.13.1
kombu==5.6.2
lgpio==0.2.2.0
libarchive-c==5.1
libclang==18.1.1
libevdev==0.11
linkify-it-py==2.0.3
lxml==5.4.0
Markdown==3.10
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
meson==1.7.0
ml_dtypes==0.5.4
more-itertools==10.7.0
namex==0.1.0
numpy==2.2.4
oauthlib==3.2.2
olefile==0.47
opencv-python==4.12.0.88
OpenEXR==1.3.10
opt_einsum==3.4.0
optree==0.18.0
packaging==25.0
paho-mqtt==1.6.1
pexpect==4.9.0
pgzero==1.2.1
picamera2==0.3.16
pidng==4.0.9
piexif==1.1.3
pigpio==1.78
pillow==11.1.0
prompt_toolkit==3.0.52
protobuf==6.33.4
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pyasyncore==1.0.2
pycairo==1.27.0
pycryptodomex==3.20.0
pycups==2.0.4
pygame==2.6.1
Pygments==2.18.0
PyGObject==3.50.0
pyinotify==0.9.6
PyJWT==2.10.1
PyOpenGL==3.1.9
pyOpenSSL==25.0.0
PyQt5==5.15.11
PyQt5_sip==12.17.0
pyserial==3.5
pysmbc==1.0.25.1
python-apt==3.0.0
python-dateutil==2.9.0
python-decouple==3.8
python-dotenv==1.0.1
python-prctl==1.8.1
pytz==2024.1
pyudev==0.24.3
PyYAML==6.0.2
redis==5.0.1
referencing==0.36.2
reportlab==4.3.1
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987==1.3.8
rgpio==0.2.2.0
rich==13.9.4
rlPyCairo==0.3.0
rpds-py==0.21.0
rpi-keyboard-config==1.0
rpi-lgpio==0.6
RTIMULib==7.2.1
sense-hat==2.6.0
setuptools==80.9.0
simplejpeg==1.8.1
six==1.17.0
smbus==1.1
smbus2==0.4.3
soupsieve==2.7
spidev==3.6
sqlparse==0.5.5
ssh-import-id==5.10
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.3.0
tqdm==4.67.1
typeguard==4.4.2
types-aiofiles==24.1
types-antlr4-python3-runtime==4.13
types-assertpy==1.1
types-atheris==2.3
types-aws-xray-sdk==2.14
types-beautifulsoup4==4.12
types-bleach==6.2
types-boltons==24.1
types-braintree==4.31
types-cachetools==5.5
types-caldav==1.3
types-capturer==3.0
types-cffi==1.16
types-chevron==0.14
types-click-default-group==1.2
types-click-spinner==0.1
types-colorama==0.4
types-commonmark==0.9
types-console-menu==0.8
types-corus==0.10
types-croniter==5.0.1
types-dateparser==1.2
types-decorator==5.1
types-defusedxml==0.7
types-Deprecated==1.2.15
types-docker==7.1
types-dockerfile-parse==2.0
types-docutils==0.21
types-editdistance==0.8
types-entrypoints==0.4
types-ExifRead==3.0
types-fanstatic==1.4
types-first==2.0
types-flake8==7.1
types-flake8-bugbear==24.12.12
types-flake8-builtins==2.5
types-flake8-docstrings==1.7
types-flake8-rst-docstrings==0.3
types-flake8-simplify==0.21
types-flake8-typing-imports==1.16
types-Flask-Cors==5.0
types-Flask-Migrate==4.0
types-Flask-SocketIO==5.4
types-fpdf2==2.8.2
types-gdb==15.0
types-gevent==24.11
types-google-cloud-ndb==2.3
types-greenlet==3.1
types-hdbcli==2.23
types-html5lib==1.1
types-httplib2==0.22
types-humanfriendly==10.0
types-hvac==2.3
types-ibm-db==3.2.4
types-icalendar==6.1
types-influxdb-client==1.45
types-inifile==0.4
types-JACK-Client==0.5
types-Jetson.GPIO==2.1
types-jmespath==1.0
types-jsonschema==4.23
types-jwcrypto==1.5
types-keyboard==0.13
types-ldap3==2.9
types-libsass==0.23
types-lupa==2.2
types-lzstring==1.0
types-m3u8==6.0
types-Markdown==3.7
types-mock==5.1
types-mypy-extensions==1.0
types-mysqlclient==2.2
types-nanoid==2.0.0
types-netaddr==1.3
types-netifaces==0.11
types-networkx==3.4.2
types-oauthlib==3.2
types-objgraph==3.6
types-olefile==0.47
types-openpyxl==3.1.5
types-opentracing==2.4
types-paramiko==3.5
types-parsimonious==0.10
types-passlib==1.7
types-passpy==1.0
types-peewee==3.17.8
types-pep8-naming==0.14
types-pexpect==4.9
types-pika-ts==1.3
types-polib==1.2
types-portpicker==1.6
types-protobuf==5.29.1
types-psutil==6.1
types-psycopg2==2.9.10
types-pyasn1==0.6
types-pyaudio==0.2
types-PyAutoGUI==0.9
types-pycocotools==2.0
types-pycurl==7.45.4
types-pyfarmhash==0.4
types-pyflakes==3.2
types-pygit2==1.15
types-Pygments==2.18
types-pyinstaller==6.11
types-pyjks==20.0
types-PyMySQL==1.1
types-pynput==1.7.7
types-pyOpenSSL==24.1
types-pyRFC3339==2.0.1
types-PyScreeze==1.0.1
types-pyserial==3.5
types-pysftp==0.2
types-pytest-lazy-fixture==0.6
types-python-crontab==3.2
types-python-datemath==3.0.1
types-python-dateutil==2.9
types-python-http-client==3.3.7
types-python-jenkins==1.8
types-python-jose==3.3
types-python-nmap==0.7
types-python-xlib==0.33
types-pytz==2024.2
types-pywin32==308
types-pyxdg==0.28
types-PyYAML==6.0
types-qrbill==1.1
types-qrcode==8.0
types-regex==2024.11.6
types-reportlab==4.2.5
types-requests==2.32
types-requests-oauthlib==2.0
types-retry==0.9
types-RPi.GPIO==0.7
types-s2clientprotocol==5
types-seaborn==0.13.2
types-Send2Trash==1.8
types-setuptools==75.6
types-shapely==2.0
types-simplejson==3.19
types-singledispatch==4.1
types-six==1.17
types-slumber==0.7
types-str2bool==1.1
types-tabulate==0.9
types-tensorflow==2.18.0
types-TgCrypto==1.2
types-toml==0.10
types-toposort==1.10
types-tqdm==4.67
types-translationstring==1.4
types-tree-sitter-languages==1.10
types-ttkthemes==3.2
types-ujson==5.10
types-unidiff==0.7
types-untangle==1.2
types-usersettings==1.1
types-uWSGI==2.0
types-vobject==0.9.9
types-waitress==3.0.1
types-WebOb==1.8
types-whatthepatch==1.0
types-workalendar==17.0
types-WTForms==3.2.1
types-wurlitzer==3.1
types-xdgenvpy==3.0
types-xmltodict==0.14
types-zstd==1.5
types-zxcvbn==4.4
typing_extensions==4.13.2
tzdata==2025.3
uc-micro-py==1.0.3
uritemplate==4.1.1
urllib3==2.3.0
v4l2-python3==0.3.5
videodev2==0.0.4
vine==5.1.0
wcwidth==0.2.14
webcolors==1.13
webencodings==0.5.1
Werkzeug==3.1.3
wheel==0.46.1
whitenoise==6.6.0
wrapt==2.0.1
zipp==3.21.0
